requirements.txt
boto3
numpy
pandas
seaborn
matplotlib
pymysql
mysql-connector-python
cryptography
joblib
xgboost==2.1.4
scikit-learn==1.6.1
mlflow==2.21.2
mlflow[extras]
kfp==2.12.1
kfp-kubernetes==1.4.0
model-registry==0.2.15
kserve==0.14
evidently==0.6.6
tabulate